requirements.txt
google-generativeai
python-dotenv
diffusers
torch
transformers
accelerate
flask